+++ spec/services/projects_for_export_service_spec.rb
@@ -44,46 +44,6 @@ RSpec.describe ProjectsForExportService do
     end
   end
 
-  describe "#transfer_by_month_projects" do
-    it "returns only transfer projects transferring in the supplied month & year" do
-      matching_project_1 = create(:transfer_project, significant_date_provisional: false, significant_date: Date.parse("2023-1-1"))
-      matching_project_2 = create(:transfer_project, significant_date_provisional: false, significant_date: Date.parse("2023-1-1"))
-      mismatching_project_1 = create(:transfer_project, significant_date_provisional: false, significant_date: Date.parse("2023-2-1"))
-      mismatching_project_2 = create(:conversion_project, significant_date_provisional: false, significant_date: Date.parse("2023-1-1"))
-
-      projects_for_export = described_class.new.transfer_projects_by_significant_date(month: 1, year: 2023)
-
-      expect(projects_for_export).to include(matching_project_1, matching_project_2)
-      expect(projects_for_export).not_to include(mismatching_project_1, mismatching_project_2)
-    end
-
-    it "includes both provisional and confirmed projects" do
-      confirmed_project = create(:transfer_project, transfer_date_provisional: false, significant_date: Date.parse("2023-1-1"))
-      provisional_project = create(:transfer_project, transfer_date_provisional: true, significant_date: Date.parse("2023-1-1"))
-
-      projects_for_export = described_class.new.transfer_projects_by_significant_date(month: 1, year: 2023)
-
-      expect(projects_for_export).to include(confirmed_project)
-      expect(projects_for_export).to include(provisional_project)
-    end
-
-    it "includes Form a MAT transfers" do
-      project = create(:transfer_project, :form_a_mat, significant_date_provisional: false, significant_date: Date.parse("2023-1-1"))
-
-      projects_for_export = described_class.new.transfer_projects_by_significant_date(month: 1, year: 2023)
-
-      expect(projects_for_export).to include(project)
-    end
-
-    it "does not include deleted projects" do
-      project = create(:transfer_project, :deleted, significant_date_provisional: false, significant_date: Date.parse("2023-1-1"))
-
-      projects_for_export = described_class.new.transfer_projects_by_significant_date(month: 1, year: 2023)
-
-      expect(projects_for_export).to_not include(project)
-    end
-  end
-
   def mock_academies_api_client_get_establishments_and_trusts
     api_client = Api::AcademiesApi::Client.new
     allow(Api::AcademiesApi::Client).to receive(:new).and_return(api_client)